requirements.txt
paho-mqtt>=1.4.0
flask>=0.12
